[IMP] some cleanups in graph/kanban views
notable: disable the group by integration in graph view, because of crash (the search view was rewritten and too much search logic was in the graph view.  This is a temporary measure, the graph view will be split into pivot/graph views, and the search integration will be done correctly in the near future.