[FIX] catch possible exception in search.js
The search view tries to evaluate the category of a filter in order
to put it in the filters or in the group by sub menu.  The pyeval.eval
function could throw an exception, which is not cool if it crashes the
client.  This fix make sure that the exception is caught.